refactor: GraphEditor custom hooks 抽出 + NodeStyle スキーマ構造化
[HIGH] GraphEditor.tsx の責務過多を解消 (custom hooks 抽出)
- useGroupNodes: グループ化ロジック + Cmd+G ショートカット
- useClipboard: コピー/ペーストロジック + Cmd+C/V ショートカット
- useEdgeContextMenu: エッジコンテキストメニュー state と handlers
- usePaneDoubleClick: ダブルクリックによるノード追加検出
- GraphEditorInner: ~500行 → ~200行, hook 呼び出しと ReactFlow への
  prop 渡しに集中

[HIGH] NodeStyle スキーマの構造化 (型安全性向上)
- StyleSchema (z.record) から NodeStyleSchema (.catchall) に移行
  x, y: z.number() / width, height: z.number()|z.string() / nodeType: z.literal('group')
- graphTransform.toFlowNodes の型ガード削除:
  typeof n.style?.x === 'number' → n.style?.x ?? 0
- 既存 JSON との後方互換性は .catchall(z.unknown()) で確保

Confidence: high
Scope-risk: moderate
Not-tested: hooks の単体テスト (graphTransform の既存47件は全パス)
